Extract Glue metadata to s3._read
