+++ awswrangler/s3/_read_orc.py
@@ -8,8 +8,6 @@ from typing import (
     Any,
     Callable,
     Dict,
-    Iterable,
-    Iterator,
     List,
     Optional,
     Union,
@@ -27,8 +25,6 @@ from awswrangler._arrow import _add_table_partitions
 from awswrangler._config import apply_configs
 from awswrangler._distributed import engine
 from awswrangler._executor import _BaseExecutor, _get_executor
-from awswrangler.catalog._get import _get_partitions
-from awswrangler.catalog._utils import _catalog_id
 from awswrangler.distributed.ray import ray_get
 from awswrangler.s3._fs import open_s3_object
 from awswrangler.s3._list import _path2list
@@ -39,6 +35,7 @@ from awswrangler.s3._read import (
     _extract_partitions_metadata_from_paths,
     _get_path_ignore_suffix,
     _get_path_root,
+    _get_paths_for_glue_table,
     _InternalReadTableMetadataReturnValue,
     _TableMetadataReader,
     _validate_schemas_from_files,
@@ -48,22 +45,10 @@ from awswrangler.typing import RaySettings, _ReadTableMetadataReturnValue
 if TYPE_CHECKING:
     from mypy_boto3_s3 import S3Client
 
-BATCH_READ_BLOCK_SIZE = 65_536
-CHUNKED_READ_S3_BLOCK_SIZE = 10_485_760  # 10 MB (20 * 2**20)
 FULL_READ_S3_BLOCK_SIZE = 20_971_520  # 20 MB (20 * 2**20)
 METADATA_READ_S3_BLOCK_SIZE = 131_072  # 128 KB (128 * 2**10)
 
 _logger: logging.Logger = logging.getLogger(__name__)
-
-
-def _ensure_locations_are_valid(paths: Iterable[str]) -> Iterator[str]:
-    for path in paths:
-        suffix: str = path.rpartition("/")[2]
-        # If the suffix looks like a partition,
-        if suffix and (suffix.count("=") == 1):
-            # the path should end in a '/' character.
-            path = f"{path}/"  # ruff: noqa: PLW2901
-        yield path
 
 
 def _pyarrow_orc_file_wrapper(source: Any) -> pyarrow.orc.ORCFile:
@@ -517,40 +502,19 @@ def read_orc_table(
     >>> df = wr.s3.read_orc_table(path, dataset=True, partition_filter=my_filter)
 
     """
-    client_glue = _utils.client(service_name="glue", session=boto3_session)
-    s3_client = _utils.client(service_name="s3", session=boto3_session)
-    res = client_glue.get_table(**_catalog_id(catalog_id=catalog_id, DatabaseName=database, Name=table))
-    try:
-        location: str = res["Table"]["StorageDescriptor"]["Location"]
-        path: str = location if location.endswith("/") else f"{location}/"
-    except KeyError as ex:
-        raise exceptions.InvalidTable(f"Missing s3 location for {database}.{table}.") from ex
-    path_root: Optional[str] = None
-    paths: Union[str, List[str]] = path
-    # If filter is available, fetch & filter out partitions
-    # Then list objects & process individual object keys under path_root
-    if partition_filter:
-        available_partitions_dict = _get_partitions(
-            database=database,
-            table=table,
-            catalog_id=catalog_id,
-            boto3_session=boto3_session,
-        )
-        available_partitions = list(_ensure_locations_are_valid(available_partitions_dict.keys()))
-        if available_partitions:
-            paths = []
-            path_root = path
-            partitions: Union[str, List[str]] = _apply_partition_filter(
-                path_root=path_root, paths=available_partitions, filter_func=partition_filter
-            )
-            for partition in partitions:
-                paths += _path2list(
-                    path=partition,
-                    s3_client=s3_client,
-                    suffix=filename_suffix,
-                    ignore_suffix=_get_path_ignore_suffix(path_ignore_suffix=filename_ignore_suffix),
-                    s3_additional_kwargs=s3_additional_kwargs,
-                )
+    paths: Union[str, List[str]]
+    path_root: Optional[str]
+
+    paths, path_root, res = _get_paths_for_glue_table(
+        table=table,
+        database=database,
+        filename_suffix=filename_suffix,
+        filename_ignore_suffix=filename_ignore_suffix,
+        catalog_id=catalog_id,
+        partition_filter=partition_filter,
+        boto3_session=boto3_session,
+        s3_additional_kwargs=s3_additional_kwargs,
+    )
 
     df = read_orc(
         path=paths,
